requirements.txt
tensorflow-cpu
flask
gunicorn